fix(#1588): move agent_registry.py + agent_record.py from core/ to services/agents/
AgentRegistry belongs in services/ per KERNEL-ARCHITECTURE.md §3 — it uses
SQLAlchemy session_factory, manages lifecycle state machines, and is not a
kernel primitive. Move both files and update all import paths.
