+++ src/nexus/services/agents/agent_record.py
