+++ src/nexus/services/agents/async_agent_registry.py
@@ -16,12 +16,12 @@ from __future__ import annotations
 import asyncio
 from typing import TYPE_CHECKING, Any
 
-from nexus.core.agent_record import AgentState
+from nexus.services.agents.agent_record import AgentState
 from nexus.services.protocols.agent_registry import AgentInfo
 
 if TYPE_CHECKING:
-    from nexus.core.agent_record import AgentRecord
-    from nexus.core.agent_registry import AgentRegistry
+    from nexus.services.agents.agent_record import AgentRecord
+    from nexus.services.agents.agent_registry import AgentRegistry
 
 
 def _to_agent_info(record: AgentRecord) -> AgentInfo:
