+++ src/nexus/services/delegation/service.py
@@ -31,7 +31,7 @@ from nexus.services.delegation.models import DelegationMode, DelegationRecord, D
 if TYPE_CHECKING:
     from sqlalchemy.orm import Session, sessionmaker
 
-    from nexus.core.agent_registry import AgentRegistry
+    from nexus.services.agents.agent_registry import AgentRegistry
     from nexus.services.permissions.entity_registry import EntityRegistry
     from nexus.services.permissions.namespace_manager import NamespaceManager
     from nexus.services.permissions.rebac_manager_enhanced import EnhancedReBACManager
