+++ tests/e2e/test_agent_registry_e2e.py
@@ -32,8 +32,8 @@ import pytest
 from sqlalchemy import create_engine, text
 from sqlalchemy.orm import sessionmaker
 
-from nexus.core.agent_record import AgentState
-from nexus.core.agent_registry import (
+from nexus.services.agents.agent_record import AgentState
+from nexus.services.agents.agent_registry import (
     AgentRegistry,
     InvalidTransitionError,
     StaleAgentError,
