+++ tests/e2e/test_async_wrappers_e2e.py
@@ -25,10 +25,10 @@ import pytest
 from sqlalchemy import create_engine
 from sqlalchemy.orm import sessionmaker
 
-from nexus.core.agent_registry import AgentRegistry, InvalidTransitionError
 from nexus.core.router import PathNotMountedError, PathRouter
 from nexus.plugins.async_hooks import AsyncHookEngine
 from nexus.plugins.hooks import PluginHooks
+from nexus.services.agents.agent_registry import AgentRegistry, InvalidTransitionError
 from nexus.services.agents.async_agent_registry import AsyncAgentRegistry
 from nexus.services.permissions.async_namespace_manager import AsyncNamespaceManager
 from nexus.services.protocols.agent_registry import AgentInfo, AgentRegistryProtocol
@@ -420,8 +420,8 @@ class TestServerLifespanWiring:
     @pytest.mark.asyncio()
     async def test_lifespan_wiring_with_real_db(self, tmp_path: Path) -> None:
         """Simulate server lifespan: AgentRegistry + AsyncAgentRegistry wiring."""
-        from nexus.core.agent_registry import AgentRegistry
         from nexus.server import fastapi_server
+        from nexus.services.agents.agent_registry import AgentRegistry
         from nexus.services.agents.async_agent_registry import AsyncAgentRegistry
 
         # Create real SQLite-backed AgentRegistry (same as server lifespan does)
@@ -481,11 +481,11 @@ class TestServerLifespanWiring:
         from unittest.mock import MagicMock
 
         from nexus.backends.local import LocalBackend
-        from nexus.core.agent_registry import AgentRegistry
         from nexus.core.protocols.vfs_router import VFSRouterProtocol
         from nexus.core.router import PathRouter
         from nexus.plugins.async_hooks import AsyncHookEngine
         from nexus.plugins.hooks import PluginHooks
+        from nexus.services.agents.agent_registry import AgentRegistry
         from nexus.services.agents.async_agent_registry import AsyncAgentRegistry
         from nexus.services.permissions.async_namespace_manager import AsyncNamespaceManager
         from nexus.services.protocols.hook_engine import HookEngineProtocol
@@ -562,7 +562,7 @@ class TestServerLifespanWiring:
         agent info. This test verifies the async wrapper provides correct data
         for permission decisions.
         """
-        from nexus.core.agent_registry import AgentRegistry
+        from nexus.services.agents.agent_registry import AgentRegistry
         from nexus.services.agents.async_agent_registry import AsyncAgentRegistry
 
         db_path = tmp_path / f"perm_{uuid.uuid4().hex[:8]}.db"
