+++ tests/integration/core/test_agent_registry_e2e.py
@@ -21,8 +21,8 @@ from sqlalchemy import create_engine
 from sqlalchemy.orm import sessionmaker
 from sqlalchemy.pool import StaticPool
 
-from nexus.core.agent_record import AgentState
-from nexus.core.agent_registry import AgentRegistry
+from nexus.services.agents.agent_record import AgentState
+from nexus.services.agents.agent_registry import AgentRegistry
 from nexus.services.delegation.models import DelegationMode
 from nexus.services.delegation.service import DelegationService
 from nexus.services.permissions.entity_registry import EntityRegistry
