+++ tests/integration/core/test_agent_registry_server.py
@@ -17,12 +17,12 @@ from sqlalchemy import create_engine
 from sqlalchemy.orm import sessionmaker
 from sqlalchemy.pool import StaticPool
 
-from nexus.core.agent_registry import AgentRegistry
 from nexus.server.api.v2.routers.delegation import (
     DelegateRequest,
     DelegateResponse,
     _handle_delegation_error,
 )
+from nexus.services.agents.agent_registry import AgentRegistry
 from nexus.services.delegation.models import DelegationMode
 from nexus.services.delegation.service import DelegationService
 from nexus.services.permissions.entity_registry import EntityRegistry
