+++ tests/integration/core/test_nexus_agent_registry.py
@@ -12,8 +12,8 @@ from sqlalchemy import create_engine
 from sqlalchemy.orm import sessionmaker
 from sqlalchemy.pool import StaticPool
 
-from nexus.core.agent_record import AgentState
-from nexus.core.agent_registry import AgentRegistry
+from nexus.services.agents.agent_record import AgentState
+from nexus.services.agents.agent_registry import AgentRegistry
 from nexus.services.permissions.entity_registry import EntityRegistry
 from nexus.storage.models import Base
