+++ tests/integration/delegation/test_delegation_full_e2e.py
@@ -19,7 +19,7 @@ from sqlalchemy import create_engine
 from sqlalchemy.orm import sessionmaker
 from sqlalchemy.pool import StaticPool
 
-from nexus.core.agent_registry import AgentRegistry
+from nexus.services.agents.agent_registry import AgentRegistry
 from nexus.services.delegation.errors import DelegationChainError, EscalationError
 from nexus.services.delegation.models import DelegationMode
 from nexus.services.delegation.service import DelegationService
