+++ tests/unit/core/test_agent_registry.py
@@ -20,8 +20,8 @@ from sqlalchemy import create_engine
 from sqlalchemy.orm import sessionmaker
 from sqlalchemy.pool import StaticPool
 
-from nexus.core.agent_record import AgentState
-from nexus.core.agent_registry import (
+from nexus.services.agents.agent_record import AgentState
+from nexus.services.agents.agent_registry import (
     AgentRegistry,
     InvalidTransitionError,
     StaleAgentError,
