+++ tests/unit/sandbox/test_auth_service.py
@@ -24,10 +24,10 @@ from sqlalchemy import create_engine
 from sqlalchemy.orm import sessionmaker
 from sqlalchemy.pool import StaticPool
 
-from nexus.core.agent_record import AgentRecord, AgentState
-from nexus.core.agent_registry import AgentRegistry, InvalidTransitionError
 from nexus.sandbox.auth_service import SandboxAuthService
 from nexus.sandbox.events import AgentEventLog
+from nexus.services.agents.agent_record import AgentRecord, AgentState
+from nexus.services.agents.agent_registry import AgentRegistry, InvalidTransitionError
 from nexus.storage.models import Base
 
 # ---------------------------------------------------------------------------
